requirements.txt
requests==2.31.0
pandas==2.2.0
pyarrow==15.0.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.0
python-dotenv==1.0.0
